+++ aio/tools/transforms/angular-base-package/index.js
@@ -36,10 +36,10 @@ module.exports = new Package('angular-base', [
   .factory('packageInfo', function() { return require(path.resolve(PROJECT_ROOT, 'package.json')); })
   .factory(require('./readers/json'))
   .factory(require('./services/copyFolder'))
-  .factory(require('./services/filterPipes'))
-  .factory(require('./services/filterAmbiguousDirectiveAliases'))
-  .factory(require('./services/ignoreHttpInUrls'))
   .factory(require('./services/getImageDimensions'))
+  .factory(require('./services/auto-link-filters/filterPipes'))
+  .factory(require('./services/auto-link-filters/filterAmbiguousDirectiveAliases'))
+  .factory(require('./services/auto-link-filters/ignoreHttpInUrls'))
 
   .factory(require('./post-processors/add-image-dimensions'))
   .factory(require('./post-processors/auto-link-code'))
