+++ aio/tools/transforms/angular-base-package/services/auto-link-filters/ignoreHttpInUrls.js
